requirements.txt
flask==2.3.3
flask-cors==4.0.0

# Networking
httpx>=0.27.0

# Audio / TTS
piper-tts
pyaudio
numpy

# ASR
RealtimeSTT

# (Optional) add any extras below
# uvloop ; platform_system=='Linux'
